pyproject.toml
[project]
name = "databricks-with-claude"
version = "0.1.0"
description = "PySpark ETL modules for Databricks"
requires-python = ">=3.12"
dependencies = [
    "databricks-connect>=17",
    "pip>=25.2",
    "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.15.1",
]
